[Mode][GitHub] Add initial implementation of the triggers. (#13)
Provide the initial work for the GitHub actions model. At the moment
we have just the Workflow (is a pipeline in the AzDO world) triggers.

GitHub actions provides triggers for all its webhooks and a few manual
ones as well as a cron job. Tests have been added monstly to show the
API and make sure we are not doing anything stupid.