Update Step template to add a task retry property
An option was added to ADO tasks in 2021 to enable immediate retry of the task on failure. This allows for failures due to intermittency or network latency etc. to be retried without a user having to re-run the job from the UI.  This change adds the option of specifying retryCountOnTaskFailure as a string value, as specified here https://learn.microsoft.com/en-us/azure/devops/pipelines/yaml-schema/steps-task?view=azure-pipelines