Fix: RelativeToGitRoot for git worktrees. (#442)
This allows Sharpliner to use the RelativeToGitRoot Path Type [on a Git
Worktree](https://git-scm.com/docs/git-worktree). Git Worktrees do not
contain a .git Directory, however, they contain a .git file, which is
used as a sort of "pointer" file to the main checkout. Looking for both
this file as well as the directory covers both scenarios.